MyProjects/s-dia-main/requirements.txt
PySide6>=6.5.0
pyinstaller>=6.3.0
pywin32>=306
psutil>=5.9.0
cryptography>=43.0.1
pycryptodome>=3.18.0
aria2p>=0.12.1
